move cosmwasm to its own module section
